+++ NEWS.md
@@ -1,3 +1,13 @@
+# onehot 0.1.3
+
+* Fixed examples in the README
+
+# onehot 0.1.2
+
+* Refactored how missing values are predicted.
+* Factors can have NA indicator columns.
+* Numerics instead use sentinel fields.
+
 # onehot 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.